+++ src/module/lib/lib.ts
@@ -126,7 +126,7 @@ export async function renderAutomatedPolymorpherHud(app, html, hudToken) {
     return;
   }
 
-  const actor = <Actor>sourceToken.actor;
+  const actor = retrieveActorFromToken(sourceToken);
   if (!actor) {
     // warn(`No actor founded on canvas with token '${sourceToken.id}'`, true);
     return;
@@ -154,13 +154,12 @@ function addToPolymorphButton(html, sourceToken: Token) {
     return;
   }
 
-  // const isPolymorphed = sourceToken.actor?.getFlag('dnd5e', 'isPolymorphed');
   const button = buildButton(html, `Transform ${sourceToken.name}`);
   // if (isPolymorphed) {
   //   button = addSlash(button);
   // }
 
-  const actor = <Actor>sourceToken.actor;
+  const actor = retrieveActorFromToken(sourceToken);
   if (!actor) {
     // warn(`No actor founded on canvas with token '${sourceToken.id}'`, true);
     return;
@@ -168,6 +167,7 @@ function addToPolymorphButton(html, sourceToken: Token) {
 
   const random = <boolean>actor?.getFlag(CONSTANTS.MODULE_NAME, PolymorpherFlags.RANDOM) ?? false;
   const ordered = <boolean>actor?.getFlag(CONSTANTS.MODULE_NAME, PolymorpherFlags.ORDERED) ?? false;
+  const storeonactor = <boolean>actor?.getFlag(CONSTANTS.MODULE_NAME, PolymorpherFlags.STORE_ON_ACTOR) ?? false;
 
   button.find('i').on('click', async (ev) => {
     for (const targetToken of <Token[]>canvas.tokens?.controlled) {
@@ -266,4 +266,52 @@ function addSlash(button) {
 function removeSlash(button) {
   const slash = button.find('i')[1];
   slash.remove();
+}
+
+
+export function retrieveActorFromToken(sourceToken:Token):Actor|undefined{
+  if(!sourceToken.actor){
+    return undefined;
+  }
+  const storeOnActorFlag = <boolean>sourceToken.actor.getFlag(CONSTANTS.MODULE_NAME, PolymorpherFlags.STORE_ON_ACTOR);
+  if(!storeOnActorFlag){
+    return sourceToken.actor;
+  }
+  let actor:Actor|undefined = undefined;
+  if(sourceToken.data.actorLink){
+    actor = <Actor>game.actors?.get(<string>sourceToken.data.actorId);
+  }
+  // DO NOT NEED THIS
+  // if(!actor){
+  //   actor = <Actor>game.actors?.get(<string>sourceToken.actor?.id);
+  // }
+  if(!actor){
+    actor = sourceToken.actor;
+  }
+  return actor;
+}
+
+export async function retrieveActorFromData(aId:string, aName:string, currentCompendium:string):Promise<Actor|null>{
+  let actorToTransformLi:Actor|null = null;
+  if(currentCompendium &&
+    currentCompendium != 'none' &&
+    currentCompendium != 'nonenodelete'){
+    const pack = game.packs.get(currentCompendium);
+    if(pack){
+      await pack.getIndex();
+      for(const entityComp of pack.index){
+        const actorComp = <Actor>await pack.getDocument(entityComp._id);
+        if(actorComp.id === aId || actorComp.name === aName){
+          actorToTransformLi = actorComp;
+          break;
+        }
+      }
+    }
+  }
+  if(!actorToTransformLi){
+    actorToTransformLi = <Actor>game.actors?.contents.find((a) => {
+      return a.id === aId || a.name === aName;
+    });
+  }
+  return actorToTransformLi;
 }